refactor: extract some code from ImgurPlugin class
